Adapting products part with new backend

+++ app/src/main/java/mk/ukim/finki/eshop/adapters/OrderProductsAdapter.kt
@@ -40,7 +40,7 @@ class OrderProductsAdapter: RecyclerView.Adapter<OrderProductsAdapter.MyViewHold
         return cartItems.size
     }
 
-    fun getProduct(position: Int): Int {
+    fun getProduct(position: Int): Long {
         return cartItems[position].product.id
     }
